refactor: implement more utils
