+++ lib/Server.js
@@ -12,31 +12,27 @@
 const fs = require('fs');
 const path = require('path');
 
-const ip = require('ip');
 const tls = require('tls');
-const url = require('url');
 const http = require('http');
 const https = require('https');
 const spdy = require('spdy');
 const sockjs = require('sockjs');
 
 const killable = require('killable');
-
-const del = require('del');
 const chokidar = require('chokidar');
-
 const express = require('express');
 
 const compress = require('compression');
 const serveIndex = require('serve-index');
 const httpProxyMiddleware = require('http-proxy-middleware');
 const historyApiFallback = require('connect-history-api-fallback');
-
-const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 
+const addProgressPlugin = require('./utils/addProgressPlugin');
+const checkHost = require('./utils/checkHost');
 const createLogger = require('./utils/createLogger');
-const createCertificate = require('./utils/createCertificate');
+const getSSLCertificate = require('./utils/getSSLCertificate');
+const statsOptions = require('./utils/statsOptions');
 
 const validateOptions = require('schema-utils');
 const schema = require('./options.json');
@@ -50,15 +46,6 @@ const version = parseFloat(process.version.slice(1));
 if (version >= 8.6 && version < 10) {
   tls.DEFAULT_ECDH_CURVE = 'auto';
 }
-
-const STATS = {
-  all: false,
-  hash: true,
-  assets: true,
-  warnings: true,
-  errors: true,
-  errorDetails: false
-};
 
 function Server (compiler, options = {}, _log) {
   this.log = _log || createLogger(options);
@@ -76,6 +63,7 @@ function Server (compiler, options = {}, _log) {
   this.clientOverlay = options.overlay;
   this.clientLogLevel = options.clientLogLevel;
 
+  // Require for `checkHost`, remove in next major release
   this.publicHost = options.public;
   this.allowedHosts = options.allowedHosts;
   this.disableHostCheck = !!options.disableHostCheck;
@@ -91,23 +79,7 @@ function Server (compiler, options = {}, _log) {
   };
 
   if (this.progress) {
-    const progressPlugin = new webpack.ProgressPlugin(
-      (percent, msg, addInfo) => {
-        percent = Math.floor(percent * 100);
-
-        if (percent === 100) {
-          msg = 'Compilation completed';
-        }
-
-        if (addInfo) {
-          msg = `${msg} (${addInfo})`;
-        }
-
-        this.sockWrite(this.sockets, 'progress-update', { percent, msg });
-      }
-    );
-
-    progressPlugin.apply(compiler);
+    addProgressPlugin(compiler, this);
   }
 
   const addHooks = (compiler) => {
@@ -116,7 +88,7 @@ function Server (compiler, options = {}, _log) {
     compile.tap('webpack-dev-server', invalidPlugin);
     invalid.tap('webpack-dev-server', invalidPlugin);
     done.tap('webpack-dev-server', (stats) => {
-      this._sendStats(this.sockets, stats.toJson(STATS));
+      this._sendStats(this.sockets, stats.toJson(statsOptions));
       this._stats = stats;
     });
   };
@@ -139,10 +111,11 @@ function Server (compiler, options = {}, _log) {
     res.send('Invalid Host header');
   });
 
-  const wdmOptions = { logLevel: this.log.options.level };
-
   // middleware for serving webpack bundle
-  this.middleware = webpackDevMiddleware(compiler, Object.assign({}, options, wdmOptions));
+  this.middleware = webpackDevMiddleware(
+    compiler,
+    Object.assign({}, options, { logLevel: this.log.options.level })
+  );
 
   app.get('/__webpack_dev_server__/live.bundle.js', (req, res) => {
     res.setHeader('Content-Type', 'application/javascript');
@@ -533,45 +506,7 @@ function Server (compiler, options = {}, _log) {
     let fakeCert;
 
     if (!options.https.key || !options.https.cert) {
-      // Use a self-signed certificate if no certificate was configured.
-      // Cycle certs every 24 hours
-      const certPath = path.join(__dirname, '../ssl/server.pem');
-
-      let certExists = fs.existsSync(certPath);
-
-      if (certExists) {
-        const certTtl = 1000 * 60 * 60 * 24;
-        const certStat = fs.statSync(certPath);
-
-        const now = new Date();
-
-        // cert is more than 30 days old, kill it with fire
-        if ((now - certStat.ctime) / certTtl > 30) {
-          this.log.info('SSL Certificate is more than 30 days old. Removing.');
-
-          del.sync([certPath], { force: true });
-
-          certExists = false;
-        }
-      }
-
-      if (!certExists) {
-        this.log.info('Generating SSL Certificate');
-
-        const attrs = [
-          { name: 'commonName', value: 'localhost' }
-        ];
-
-        const pems = createCertificate(attrs);
-
-        fs.writeFileSync(
-          certPath,
-          pems.private + pems.cert,
-          { encoding: 'utf-8' }
-        );
-      }
-
-      fakeCert = fs.readFileSync(certPath);
+      fakeCert = getSSLCertificate(this);
     }
 
     options.https.key = options.https.key || fakeCert;
@@ -625,83 +560,11 @@ Server.prototype.setContentHeaders = function (req, res, next) {
   next();
 };
 
-Server.prototype.checkHost = function (headers) {
-  // allow user to opt-out this security check, at own risk
-  if (this.disableHostCheck) {
-    return true;
-  }
-  // get the Host header and extract hostname
-  // we don't care about port not matching
-  const hostHeader = headers.host;
-
-  if (!hostHeader) {
-    return false;
-  }
-
-  // use the node url-parser to retrieve the hostname from the host-header.
-  const hostname = url.parse(`//${hostHeader}`, false, true).hostname;
-  // always allow requests with explicit IPv4 or IPv6-address.
-  // A note on IPv6 addresses:
-  // hostHeader will always contain the brackets denoting
-  // an IPv6-address in URLs,
-  // these are removed from the hostname in url.parse(),
-  // so we have the pure IPv6-address in hostname.
-  if (ip.isV4Format(hostname) || ip.isV6Format(hostname)) {
-    return true;
-  }
-  // always allow localhost host, for convience
-  if (hostname === 'localhost') {
-    return true;
-  }
-  // allow if hostname is in allowedHosts
-  if (this.allowedHosts && this.allowedHosts.length) {
-    for (let hostIdx = 0; hostIdx < this.allowedHosts.length; hostIdx++) {
-      const allowedHost = this.allowedHosts[hostIdx];
-
-      if (allowedHost === hostname) return true;
-
-      // support "." as a subdomain wildcard
-      // e.g. ".example.com" will allow "example.com", "www.example.com", "subdomain.example.com", etc
-      if (allowedHost[0] === '.') {
-        // "example.com"
-        if (hostname === allowedHost.substring(1)) {
-          return true;
-        }
-        // "*.example.com"
-        if (hostname.endsWith(allowedHost)) {
-          return true;
-        }
-      }
-    }
-  }
-
-  // allow hostname of listening adress
-  if (hostname === this.hostname) {
-    return true;
-  }
-
-  // also allow public hostname if provided
-  if (typeof this.publicHost === 'string') {
-    const idxPublic = this.publicHost.indexOf(':');
-
-    const publicHostname = idxPublic >= 0
-      ? this.publicHost.substr(0, idxPublic)
-      : this.publicHost;
-
-    if (hostname === publicHostname) {
-      return true;
-    }
-  }
-
-  // disallow
-  return false;
-};
-
 // delegate listen call and init sockjs
 Server.prototype.listen = function (port, hostname, fn) {
   this.hostname = hostname;
 
-  const returnValue = this.listeningApp.listen(port, hostname, (err) => {
+  return this.listeningApp.listen(port, hostname, (err) => {
     const socket = sockjs.createServer({
       // Use provided up-to-date sockjs-client
       sockjs_url: '/__webpack_dev_server__/sockjs.bundle.js',
@@ -758,7 +621,7 @@ Server.prototype.listen = function (port, hostname, fn) {
         return;
       }
 
-      this._sendStats([ connection ], this._stats.toJson(STATS), true);
+      this._sendStats([ connection ], this._stats.toJson(statsOptions), true);
     });
 
     socket.installHandlers(this.listeningApp, {
@@ -769,8 +632,6 @@ Server.prototype.listen = function (port, hostname, fn) {
       fn.call(this.listeningApp, err);
     }
   });
-
-  return returnValue;
 };
 
 Server.prototype.close = function (cb) {
@@ -883,6 +744,9 @@ Server.prototype.invalidate = function () {
     this.middleware.invalidate();
   }
 };
+
+// Remove from public api in next major release
+Server.prototype.checkHost = checkHost;
 
 // Export this logic,
 // so that other implementations,
