+++ lib/utils/statsOptions.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+  all: false,
+  hash: true,
+  assets: true,
+  warnings: true,
+  errors: true,
+  errorDetails: false
+};
